# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend service for SmartLeague. Cookie-based auth (HttpOnly access/refresh cookies).",
        "title": "SmartLeague API",
        "contact": {
            "name": "SmartLeague Backend"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/api/v1/series/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get all series list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAllSeriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
        "/api/v1/series/{id}/full": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get series full data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "participants limit",
                        "name": "participants_limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "participants offset",
                        "name": "participants_offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "games limit",
                        "name": "games_limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "games offset",
                        "name": "games_offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "leaderboard limit",
                        "name": "leaderboard_limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "leaderboard offset",
                        "name": "leaderboard_offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetSeriesFullResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
        "/api/v1/series/{id}/join": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Join series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
        "/api/v1/series/{id}/leave": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Leave series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
        "/api/v1/series/{id}/participants": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get series participants",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetSeriesParticipantsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AllSeriesItem": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string"
                },
                "club_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "games_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.Game": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "host_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "series_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/types.GameStatus"
                }
            }
        },
...
        "dto.GetAllSeriesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AllSeriesItem"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationInfo"
                }
            }
        },
...
        "dto.GetSeriesFullResponse": {
            "type": "object",
            "properties": {
                "games": {
                    "$ref": "#/definitions/dto.GetSeriesGamesResponse"
                },
                "leaderboard": {
                    "$ref": "#/definitions/dto.GetSeriesLeaderboardResponse"
                },
                "participants": {
                    "$ref": "#/definitions/dto.GetSeriesParticipantsResponse"
                },
                "series": {
                    "$ref": "#/definitions/dto.Series"
                }
            }
        },
        "dto.GetSeriesGamesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Game"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationInfo"
                }
            }
        },
        "dto.GetSeriesLeaderboardResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.LeaderboardRow"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationInfo"
                }
            }
        },
        "dto.GetSeriesParticipantsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.User"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationInfo"
                }
            }
        },
...
        "dto.HTTPStatus": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.LeaderboardRow": {
            "type": "object",
            "properties": {
                "points": {
                    "type": "number"
                },
                "profile_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginationInfo": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_previous": {
                    "type": "boolean",
                    "example": false
                },
                "total_items": {
                    "type": "integer",
                    "example": 1250
                },
                "total_pages": {
                    "type": "integer",
                    "example": 13
                }
            }
        },
...
        "dto.Series": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "game_type": {
                    "$ref": "#/definitions/types.GameType"
                },
                "id": {
                    "type": "string"
                },
                "is_closed": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price_rub": {
                    "type": "integer"
                },
                "start_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/types.SeriesStatus"
                }
            }
        },
...
        "dto.User": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "club_state": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ClubState"
                        }
                    ],
                    "example": 0
                },
                "description": {
                    "type": "string",
                    "example": "About me"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "nickname": {
                    "type": "string",
                    "example": "mishmish"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Role"
                        }
                    ],
                    "example": 0
                },
                "show_name": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "types.ClubState": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "ClubStateNone",
                "ClubStateMember",
                "ClubStateLeader",
                "ClubStatePresident",
                "ClubStateResident"
            ]
        },
        "types.GameStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "GameStatusDraft",
                "GameStatusInProgress",
                "GameStatusFinished"
            ]
        },
        "types.GameType": {
            "type": "integer",
            "enum": [
                0
            ],
            "x-enum-varnames": [
                "GameTypeSportMafia"
            ]
        },
...
        "types.Role": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleModerator",
                "RoleAdmin",
                "RoleSuperAdmin"
            ]
        },
        "types.SeriesStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "SeriesStatusClosed",
                "SeriesStatusRegistration",
                "SeriesStatusClosedRegistration",
                "SeriesStatusGames"
            ]
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Authentication via HttpOnly cookies:\\n- `user_auth_access_token` (short-lived)\\n- `user_auth_refresh_token` (long-lived)\\n\\nProtected endpoints require valid cookies to be sent by the client.",
            "type": "apiKey",
            "name": "user_auth_access_token",
            "in": "cookie"
        }
    },
...
}